Switch to using Bio::DB::Big for big file work.
We have a new type of library available for working with Kent big files
called Bio::DB::Big. This is based on the library libBigWig. There are a
number of changes we have done as part of this switch.

- Removed BigBedAdaptor and BigWigAdaptor. These were deprecated anyway

- Removed usage of Bio::DB::BigFile from the modules

- Removed all summary methods from Bio::EnsEMBL::IO::BigFileParser into
  BigWigParser. Summaries on a BigBed in Bio::DB::BigFile was giving
  information about overlapping blocks and nothing about a BigBed
  value/score (logic is that this kind of bed file should have been a
  BigWig file)

- BigWig now uses get_stats (for mean values across bins) and
  get_intervals (for the underlying blocks in a wig file). This replaced
  bigWigSummaryArray() and bigWigIntervalQuery() respectively.

- BigBed now uses get_entries to retrieve bed blocks and replaces
  bigBedIntervalQuery()

- All methods have switched to using arrays that are returned from the
  Bio::DB::Big library versus using iterators

- AutoSQL methods now use the new AutoSQL perl parser and object
  representation

- Cleaned up most calls. Bio::DB::Big requires you to call the methods
  that make sense for that type of object so no more need for creating
  meta methods to call.

The only possible issue with these changes is the lack of summary
methods for BigBed files. Whilst the kent and Bio::DB::BigFile gave
access to routines for calculating some statistics at the moment
libBigWig does not. This can be raised as a future task to complete
since this is possible to do.