Fixed bins, no summary cache, improved docs and more tests
On reviewing the code it seems the bin calculation code was off by one. This has been amended. We have added test cases to verify this. All pass.

Summary array no longer caches. The original code pushed an array of features into the internal cache but never asked the cache if it existed. Summaries were always recalculated irrelevant of the fact that they were cached.

Feature cache is improved for the seek() method. Always reset the cache when we ask to seek. Previous code would have appended new features on the existing stack. 

A number of methods have been added to short-cut the underlying wig methods. There was no need to delegate.

Documentation has been improved with descriptions of what the methods do and the return values.